# IdentifierSet performance improvements